feat: 관리자 패키지 분리
#8
