+++ backend/src/main/java/sw_css/milestone/api/MilestoneHistoryController.java
@@ -5,7 +5,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -13,7 +12,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import sw_css.milestone.application.MilestoneHistoryCommandService;
 import sw_css.milestone.application.dto.request.MilestoneHistoryCreateRequest;
-import sw_css.milestone.application.dto.request.MilestoneHistoryRejectRequest;
 
 @Validated
 @RequestMapping("/milestones/histories")
@@ -33,21 +31,6 @@ public class MilestoneHistoryController {
     @DeleteMapping("/{historyId}")
     public ResponseEntity<Void> deleteMilestoneHistory(@PathVariable("historyId") final Long historyId) {
         milestoneHistoryCommandService.deleteMilestoneHistory(historyId);
-        return ResponseEntity.noContent().build();
-    }
-
-    // TODO 관리자만 호출할 수 있도록 권한 설정
-    @PatchMapping("/{historyId}/approve")
-    public ResponseEntity<Void> approveMilestoneHistory(@PathVariable("historyId") final Long historyId) {
-        milestoneHistoryCommandService.approveMilestoneHistory(historyId);
-        return ResponseEntity.noContent().build();
-    }
-
-    // TODO 관리자만 호출할 수 있도록 권한 설정
-    @PatchMapping("/{historyId}/reject")
-    public ResponseEntity<Void> approveMilestoneHistory(@PathVariable("historyId") final Long historyId,
-                                                        @RequestBody final MilestoneHistoryRejectRequest request) {
-        milestoneHistoryCommandService.rejectMilestoneHistory(historyId, request);
         return ResponseEntity.noContent().build();
     }
 }
