+++ backend/src/main/java/sw_css/milestone/application/MilestoneHistoryCommandService.java
@@ -4,7 +4,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import sw_css.milestone.application.dto.request.MilestoneHistoryCreateRequest;
-import sw_css.milestone.application.dto.request.MilestoneHistoryRejectRequest;
 import sw_css.milestone.domain.Milestone;
 import sw_css.milestone.domain.MilestoneHistory;
 import sw_css.milestone.domain.repository.MilestoneHistoryRepository;
@@ -39,19 +38,5 @@ public class MilestoneHistoryCommandService {
                 .orElseThrow(
                         () -> new MilestoneHistoryException(MilestoneHistoryExceptionType.NOT_FOUND_MILESTONE_HISTORY));
         history.delete();
-    }
-
-    public void approveMilestoneHistory(final Long historyId) {
-        final MilestoneHistory history = milestoneHistoryRepository.findById(historyId)
-                .orElseThrow(
-                        () -> new MilestoneHistoryException(MilestoneHistoryExceptionType.NOT_FOUND_MILESTONE_HISTORY));
-        history.approve();
-    }
-
-    public void rejectMilestoneHistory(final Long historyId, final MilestoneHistoryRejectRequest request) {
-        final MilestoneHistory history = milestoneHistoryRepository.findById(historyId)
-                .orElseThrow(
-                        () -> new MilestoneHistoryException(MilestoneHistoryExceptionType.NOT_FOUND_MILESTONE_HISTORY));
-        history.reject(request.reason());
     }
 }
