+++ backend/src/test/java/sw_css/restdocs/RestDocsTest.java
@@ -14,6 +14,7 @@ import org.springframework.restdocs.mockmvc.RestDocumentationResultHandler;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
 import org.springframework.web.filter.CharacterEncodingFilter;
+import sw_css.admin.milestone.application.MilestoneHistoryAdminCommandService;
 import sw_css.helper.ApiTestHelper;
 import sw_css.major.application.MajorQueryService;
 import sw_css.milestone.application.MilestoneHistoryCommandService;
@@ -27,6 +28,9 @@ public abstract class RestDocsTest extends ApiTestHelper {
 
     @MockBean
     protected MilestoneHistoryCommandService milestoneHistoryCommandService;
+
+    @MockBean
+    protected MilestoneHistoryAdminCommandService milestoneHistoryAdminCommandService;
 
     @Autowired
     protected RestDocumentationResultHandler restDocs;
